Chatbox colour fix
Fixes the chatbox color bug that prevents custom chat colours. Was caused by Gmod saving Color tables and normal tables as the same, but treating them differently. Meaning that editting the color table saves it as a normal table, and then fucks up when its trying to be used as a Color table